Refactor query class structure
In an effort to mirror the domain of Cypher the structure of
query and graph were rewritten. Now a query is built from clauses
and each clause will implement apropriate interfaces that
contain the functions relevant to the scope of a given clause. The most
notable difference is in the use of the match and create functions
as the scope in their lambdas is now specific to that clause and not
shared across all clauses:

```
query {
  match {
  // `this` is a MatchClause in this block
  }
}
```